+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* **Import** Refac the carousel to use @vtex/slick-components.
+
+## [0.2.1] - 2018-18-19
+
+### Added
+
+* **CI** Add integration with Travis CI
+
 ## [0.2.0] - 2018-18-04
 
 ### Added